requirements.txt
fastapi==0.70.0
python-multipart==0.0.5
uvicorn==0.15.0
jinja2==3.0.3
tensorflow==2.6.0
keras==2.6.0
pillow==8.4.0